refactor(api): add admin route and controller
for certification-center-invitation
