+++ api/src/team/application/certification-center-invitation.controller.js
@@ -1,6 +1,5 @@
 import { extractLocaleFromRequest } from '../../../lib/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../domain/usecases/index.js';
-import { certificationCenterInvitationSerializer } from '../infrastructure/serializers/jsonapi/certification-center-invitation-serializer.js';
 
 const sendInvitations = async function (request, h) {
   const certificationCenterId = request.params.certificationCenterId;
@@ -12,26 +11,6 @@ const sendInvitations = async function (request, h) {
   return h.response().code(204);
 };
 
-const sendInvitationForAdmin = async function (request, h, dependencies = { certificationCenterInvitationSerializer }) {
-  const certificationCenterId = request.params.certificationCenterId;
-  const invitationInformation = await certificationCenterInvitationSerializer.deserializeForAdmin(request.payload);
-
-  const { certificationCenterInvitation, isInvitationCreated } =
-    await usecases.createOrUpdateCertificationCenterInvitationForAdmin({
-      certificationCenterId,
-      email: invitationInformation.email,
-      locale: invitationInformation.language,
-      role: invitationInformation.role,
-    });
-
-  const serializedCertificationCenterInvitation =
-    dependencies.certificationCenterInvitationSerializer.serializeForAdmin(certificationCenterInvitation);
-  if (isInvitationCreated) {
-    return h.response(serializedCertificationCenterInvitation).created();
-  }
-  return h.response(serializedCertificationCenterInvitation);
-};
-
-const certificationCenterInvitationController = { sendInvitations, sendInvitationForAdmin };
+const certificationCenterInvitationController = { sendInvitations };
 
 export { certificationCenterInvitationController };
