+++ api/src/team/application/routes.js
@@ -1,7 +1,8 @@
+import { certificationCenterInvitationAdminRoutes } from './certification-center-invitation.admin.route.js';
 import { certificationCenterInvitationRoutes } from './certification-center-invitation.route.js';
 
 const register = async function (server) {
-  server.route(certificationCenterInvitationRoutes);
+  server.route([...certificationCenterInvitationRoutes, ...certificationCenterInvitationAdminRoutes]);
 };
 
 const name = 'team-api';
